refactor(mcp): split monolithic tools.py into modular package
- tools/linting.py: Core linting engine (_lint_content_impl, _lint_file_impl)
- tools/formatting.py: Core formatting engine (_format_content_impl, _format_file_impl)
- tools/batch_operations.py: Multi-file ops with glob support
- tools/documentation.py: Rule/formatter discovery and lookup
- tools/diagnostics.py: Statistics, suggestions, explanations
- tools/registration.py: MCP tool registration (15 async tools)
- tools/utils/: Shared constants and helper functions
