+++ src/robocop/mcp/tools/__init__.py
@@ -1,0 +1,1 @@
+"""MCP Tools for Robocop - Linting and formatting tools exposed via MCP."""
