+++ src/robocop/mcp/tools/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities for MCP tools - shared constants and helper functions."""
